+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - The Workalendar project has been moved from Peopledoc's organization to its own (#651, #653, thx to @ewjoachim).
 - Updated Japan calendar because of the Olympics, thx @lxlgarnett. (#662)
+- Fixed Japan "Sports Day" label depending on the year.
 
 ## v15.3.0 (2021-05-07)